Create Lab 02.py
the following set of data.. You should be able to inspect your In shopping for a new house, you must consider several factors. In this problem, the initial cost of the house, the estimated annual fuel cost, and the annual tax rate are available.. Write a program that will determine the total cost of a house after a five-year period for program output to determine the best buy.

Initial cost     Annual Fuel Cost Tax Rate
$167,000      $2,300     0.025
$162,000      $1,850     0.025
$175,000      $2,500     0.020

To calculate the house cost, add the initial cost to the fuel
cost for five years, then add the taxes for five years. Taxes for one year are computed by multiplying the tax rate by the initial cost. Write a program segment to display instructions
to the program user.